+++ src/Digdir.Tool.Dialogporten.LargeDataSetSeeder/DeterministicUuidV7.cs
@@ -1,6 +1,6 @@
 using UUIDNext;
 
-namespace Digdir.Tool.Dialogporten.LargeDataSetGenerator;
+namespace Digdir.Tool.Dialogporten.LargeDataSetSeeder;
 
 public static class DeterministicUuidV7
 {
