+++ src/Digdir.Tool.Dialogporten.LargeDataSetSeeder/EntityGenerators/ActorName.cs
@@ -1,8 +1,8 @@
 using System.Collections.Concurrent;
 using Npgsql;
-using static Digdir.Tool.Dialogporten.LargeDataSetGenerator.Utils;
+using static Digdir.Tool.Dialogporten.LargeDataSetSeeder.Utils;
 
-namespace Digdir.Tool.Dialogporten.LargeDataSetGenerator.EntityGenerators;
+namespace Digdir.Tool.Dialogporten.LargeDataSetSeeder.EntityGenerators;
 
 internal static class ActorName
 {
