+++ src/Digdir.Tool.Dialogporten.LargeDataSetSeeder/EntityGenerators/DialogGuiAction.cs
@@ -1,8 +1,8 @@
 using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Actions;
 using Digdir.Domain.Dialogporten.Domain.Http;
-using static Digdir.Tool.Dialogporten.LargeDataSetGenerator.Utils;
+using static Digdir.Tool.Dialogporten.LargeDataSetSeeder.Utils;
 
-namespace Digdir.Tool.Dialogporten.LargeDataSetGenerator.EntityGenerators;
+namespace Digdir.Tool.Dialogporten.LargeDataSetSeeder.EntityGenerators;
 
 internal static class DialogGuiAction
 {
