+++ src/Digdir.Tool.Dialogporten.LargeDataSetSeeder/EntityGenerators/DialogSearchTag.cs
@@ -1,6 +1,6 @@
-using static Digdir.Tool.Dialogporten.LargeDataSetGenerator.Utils;
+using static Digdir.Tool.Dialogporten.LargeDataSetSeeder.Utils;
 
-namespace Digdir.Tool.Dialogporten.LargeDataSetGenerator.EntityGenerators;
+namespace Digdir.Tool.Dialogporten.LargeDataSetSeeder.EntityGenerators;
 
 internal static class DialogSearchTag
 {
