+++ src/Digdir.Tool.Dialogporten.LargeDataSetSeeder/Words.cs
@@ -1,4 +1,4 @@
-namespace Digdir.Tool.Dialogporten.LargeDataSetGenerator;
+namespace Digdir.Tool.Dialogporten.LargeDataSetSeeder;
 
 internal static class Words
 {
